Needed to Completely Rename file for change from API to Api to be reflected in git
What
Added an 's' to the of the following files:
 - APICourseListing - APICourseListings
 - APIWebCrawler - APIWebCrawlers

Why
Wanted the file names to have the name API instead of Api

Tested
Manual test of searching and generating a course list
